+++ src/pages/BitcoinDashboard/components/PendingDeposits/PendingDeposits.module.scss
@@ -2,8 +2,6 @@
 @import 'src/styles/themes';
 
 .listpools {
-  margin-bottom: 24px;
-
   &_title {
     font-weight: 500;
     font-size: 24px;
@@ -106,6 +104,8 @@
   &_list {
     display: flex;
     gap: 35px;
+    height: 300px;
+    overflow-y: scroll;
     // margin-top: 16px;
     flex-wrap: wrap;
   }
